Dokumanlar/Flask/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-pymongo==2.3.0
python-dotenv==1.0.0
WTForms==3.0.1
pymongo==4.5.0